refactor: mock => mocks로 디렉토리 명 변경
