+++ frontend/src/mocks/drinks.ts
