+++ frontend/src/mocks/drinksDetail.ts
